week06/requirements.txt
# Week 6 Buổi 12 — Kafka Producer & Consumer
kafka-python>=2.0.2
